requirements.txt
absl-py==2.3.1
contourpy==1.3.3
cycler==0.12.1
etils==1.13.0
fonttools==4.61.1
fsspec==2025.12.0
glfw==2.10.0
imageio==2.37.2
importlib-resources==6.5.2
kiwisolver==1.4.9
matplotlib==3.10.8
mujoco==3.3.7
numpy==2.3.5
packaging==25.0
pillow==12.0.0
pyopengl==3.1.10
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pyyaml==6.0.3
six==1.17.0
trimesh==4.10.0
typing-extensions==4.15.0
zipp==3.23.0
